Respecting animation settings
If Windows is set to not show animation effects, the stars will no longer animate and the sun will disappear - its location will no longer be calculated. The current time of day remains frozen. TODO: Will just need to fix dynamic day/night switching from the weather API in the future because that will not work when the system isn't responding to cursor changes.