Prevented cloud bunching
Clouds are less likely to bunch up outside of the screen. This used to be a problem as all clouds were spawned inside of the screen and proceeded to move out, so there was a period where all had moved out but hadn't yet re-entered, inaccurately representing the coverage on-screen.