Animations smoothly halt on focus loss
When the window loses focus, the animations of the cloud and the stars (soon to be also the rain, etc.) smoothly come to a stop and stop draining resources once they have stopped altogether. Once the window regains focus, the animations will resume.

Together with the sun simulation being suspended as the sun loses kinetic energy, this helps the app have a near-zero CPU and GPU footprint while the user uses other applications on their computer, which should be especially helpful on mobile devices.